Streamlined forecasting over multiple events.
In particular, we updated forecasting so that we rely upon not CHUNKS but the list of context times (including np.nan) for each event.

This lets our forecasting be aligned with how we export data to disk.  It also lets us remove some reundnant functions with alternate data representations - a situation just asking for bugs and developmental divergence.